custom hooks - use coordinates + nearby
