+++ client/src/layouts/MainPage.jsx
@@ -6,24 +6,18 @@ import Display from "../components/Display";
 import SearchBar from "../components/SearchBar";
 import SignedOutNav from "../components/SignedOutNav";
 import Toast from "../components/Validation Toast/Toast";
+import useCoordinates from "../hooks/useCoordinates";
+import useNearby from "../hooks/useNearby";
 
 function MainPage() {
-  const [coordinates, setCoordinates] = useState({});
-  const [nearby, setNearby] = useState({});
+  // added the useCoordinates hook to extract user coordinate. 
+  const coordinates = useCoordinates();
+  // added useNearby to extract nearby data
+  const nearby = useNearby();
   const [results, setResults] = useState({});
   const [error, setError] = useState([]);
 
   const [isResults, setISResults] = useState(false);
-
-  useEffect(() => {
-    navigator.geolocation.getCurrentPosition((position) => {
-      const newCoordinates = {
-        lat: position.coords.latitude,
-        long: position.coords.longitude,
-      };
-      setCoordinates(newCoordinates);
-    });
-  }, []);
 
   useEffect(() => {
     if (Object.entries(results).length > 0) {
@@ -38,61 +32,6 @@ function MainPage() {
     // console.log('i am here', coordinates);
     localStorage.setItem("coord", JSON.stringify(coordinates));
   }, [coordinates]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const oldVal = localStorage.getItem("coord");
-
-      try {
-        const parsedVal = JSON.parse(oldVal);
-
-        if (
-          localStorage.getItem("nearby") ===undefined && localStorage.getItem("coord")
-        ) {
-          // Do nothing
-          // setNearby(parsedVal)
-          getNearbyValue(coordinates.lat, coordinates.long);
-          localStorage.setItem("coord", JSON.stringify(coordinates));
-          
-        } else if(localStorage.getItem("coord")) {
-          getNearbyValue(coordinates.lat, coordinates.long);
-          localStorage.setItem("coord", JSON.stringify(coordinates));
-        }else{
-          // Do the API call once
-          setNearby(JSON.parse(localStorage.getItem("nearby")));
-
-          // await getNearbyValue(coordinates.lat, coordinates.long);
-          // localStorage.setItem("coord", JSON.stringify(coordinates));
-        }
-      } catch (error) {
-        // Handle parsing error
-        console.error("Error parsing JSON:", error);
-      }
-    };
-
-    fetchData();
-  }, [coordinates]);
-
-  // This does not work for the city of Tulkarem!
-  const getNearbyValue = (lat, long) => {
-    // console.log(lat,long)
-    fetch(`http://localhost:8000/api/nearby/${lat},${long}/findnearby`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        localStorage.setItem("nearby", JSON.stringify(data.json));
-
-        setNearby(data.json);
-        // console.log(data);
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
 
   const getPicsUpdate = (locationId) => {
     fetch(`http://localhost:8000/api/places/${locationId}/findlocationmage`, {
